결제 API 데이터베이스 붙였으며, transaction 을 이용한 CRUD 작업중입니다.

+++ src/main/java/com/daalgae/daalgaeproject/board/dao/BoardMapper.java
@@ -5,7 +5,6 @@ import com.daalgae.daalgaeproject.board.dto.BoardDTO;
 import com.daalgae.daalgaeproject.board.dto.ReplyDTO;
 import com.daalgae.daalgaeproject.common.paging.SelectCriteria;
 import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
 
 import java.util.List;
 import java.util.Map;
@@ -16,13 +15,13 @@ public interface BoardMapper {
     int selectTotalCount(Map<String, String> searchMap);
     List<BoardDTO> selectBoardList(SelectCriteria selectCriteria);
 
-    BoardDTO selectBoardDetail(int no);
+    BoardDTO selectBoardDetail(Long no);
 
-    List<ReplyDTO> selectReplyList(int boardNo);
+    List<ReplyDTO> selectReplyList(Long boardNo);
 
     int insertReply(ReplyDTO registReply);
 
-    int deleteReply(int no);
+    int deleteReply(Long no);
 
     int insertBoard(BoardDTO board);
 
@@ -32,7 +31,7 @@ public interface BoardMapper {
 
     int insertAttachment(AttachmentDTO attachmentDTO);
 
-    int incrementBoardCount(int no);
+    int incrementBoardCount(Long no);
 
-    BoardDTO selectThumbnailDetail(int no);
+    BoardDTO selectThumbnailDetail(Long no);
 }
